refactor: separar feature welcome y auth, agregar comentarios concisos
- Crear feature welcome separado (WelcomeScreen, LoginButtons, GuestOption, AppLogo)
- Mantener feature auth solo con LoginScreen (login con email)
- Agregar comentarios concisos en AuthContext explicando flujo de autenticación
- Agregar comentarios en secure-store, layouts e index
- Actualizar imports en ReauthLoadingScreen y app/index.tsx
- Mejorar separación de responsabilidades por features
