+++ src/features/auth/components/index.ts
@@ -1,4 +1,1 @@
-export { WelcomeScreen } from './WelcomeScreen'
-export { AppLogo } from './AppLogo'
-export { LoginButtons } from './LoginButtons'
-export { GuestOption } from './GuestOption'
+export { LoginScreen } from './LoginScreen'
