+++ src/features/welcome/components/WelcomeScreen.tsx
@@ -1,10 +1,9 @@
-// Ubicación: src/features/auth/components/WelcomeScreen.tsx
-
 import { Stack } from "tamagui";
 import { AppLogo } from "./AppLogo";
 import { GuestOption } from "./GuestOption";
 import { LoginButtons } from "./LoginButtons";
 
+// Pantalla de bienvenida: primera pantalla que ven usuarios sin sesión activa
 export function WelcomeScreen() {
   return (
     <Stack
