+++ src/shared/utils/secure-store.ts
@@ -1,7 +1,9 @@
 import * as SecureStore from 'expo-secure-store';
 
+// Clave única para identificar el token en el almacenamiento seguro del dispositivo
 const TOKEN_KEY = 'auth_token';
 
+// Guarda el token de autenticación de forma segura en el dispositivo
 export const setToken = async (token: string): Promise<void> => {
   try {
     await SecureStore.setItemAsync(TOKEN_KEY, token);
@@ -11,6 +13,7 @@ export const setToken = async (token: string): Promise<void> => {
   }
 };
 
+// Recupera el token guardado desde el almacenamiento seguro
 export const getToken = async (): Promise<string | null> => {
   try {
     const token = await SecureStore.getItemAsync(TOKEN_KEY);
@@ -22,6 +25,7 @@ export const getToken = async (): Promise<string | null> => {
   }
 };
 
+// Elimina el token del almacenamiento seguro (usado en logout)
 export const deleteToken = async (): Promise<void> => {
   try {
     await SecureStore.deleteItemAsync(TOKEN_KEY);
